[Benchmark] Refactor MMMU into dataset→tasks→eval architecture and add speed metrics
