+++ benchmarks/dataset/mmmu.py
@@ -210,150 +210,3 @@ def load_mmmu_samples(max_samples: int | None = None) -> list[MMMUSample]:
 
     logger.info("Loaded %d MMMU samples", len(samples))
     return samples
-
-
-def _check_is_number(s: str) -> bool:
-    try:
-        float(s.replace(",", ""))
-        return True
-    except Exception:
-        return False
-
-
-def _normalize_str(s: str) -> list:
-    """Normalize a string for open-ended answer comparison."""
-    s = s.strip()
-    if _check_is_number(s):
-        s = s.replace(",", "")
-        try:
-            return [round(float(s), 2)]
-        except Exception:
-            return [s.lower()]
-    return [s.lower()] if len(s) > 1 else [" " + s, s + " "]
-
-
-def _extract_numbers(s: str) -> list[str]:
-    """Extract all numbers (with commas, scientific notation, etc.) from *s*."""
-    pattern_commas = r"-?\b\d{1,3}(?:,\d{3})+\b"
-    pattern_scientific = r"-?\d+(?:\.\d+)?[eE][+-]?\d+"
-    pattern_simple = r"-?(?:\d+\.\d+|\.\d+|\d+\b)(?![eE][+-]?\d+)(?![,\d])"
-    return (
-        re.findall(pattern_commas, s)
-        + re.findall(pattern_scientific, s)
-        + re.findall(pattern_simple, s)
-    )
-
-
-def parse_open_response(response: str) -> list:
-    """Extract answer candidates from an open-ended model response.
-
-    Ported from SGLang ``simple_eval_mmmu_vlm.py``.
-    """
-
-    def _get_key_subresponses(resp: str) -> list[str]:
-        resp = resp.strip().strip(".").lower()
-        subs = re.split(r"\.\s(?=[A-Z])|\n", resp)
-        indicators = [
-            "could be ",
-            "so ",
-            "is ",
-            "thus ",
-            "therefore ",
-            "final ",
-            "answer ",
-            "result ",
-        ]
-        keys: list[str] = []
-        for i, s in enumerate(subs):
-            cands = [*indicators]
-            if i == len(subs) - 1:
-                cands.append("=")
-            shortest = None
-            for ind in cands:
-                if ind in s:
-                    part = s.split(ind)[-1].strip()
-                    if not shortest or len(part) < len(shortest):
-                        shortest = part
-            if shortest and shortest not in (":", ",", ".", "!", "?", ";", "'"):
-                keys.append(shortest)
-        return keys or [resp]
-
-    key_resps = _get_key_subresponses(response)
-    pred_list = key_resps.copy()
-    for r in key_resps:
-        pred_list.extend(_extract_numbers(r))
-    out: list = []
-    for x in pred_list:
-        out.extend(_normalize_str(x))
-    return list(dict.fromkeys(out))
-
-
-def eval_open(gold, preds: list) -> bool:
-    """Check if any prediction matches the gold answer (fuzzy)."""
-    if isinstance(gold, list):
-        norm_answers: list = []
-        for ans in gold:
-            norm_answers.extend(_normalize_str(ans))
-    else:
-        norm_answers = _normalize_str(gold)
-    for p in preds:
-        if isinstance(p, str):
-            for na in norm_answers:
-                if isinstance(na, str) and na in p:
-                    return True
-        else:
-            if p in norm_answers:
-                return True
-    return False
-
-
-def parse_multi_choice_response(
-    response: str,
-    all_choices: list[str],
-    index2ans: dict[str, str],
-) -> str:
-    """Extract a single answer letter from the model response.
-
-    Ported from SGLang ``simple_eval_mmmu_vlm.py`` with the same extraction
-    priority: ``(A)`` bracket → ``·A·`` space-padded → option-text match →
-    last-occurrence tie-break → fallback to first choice.
-    """
-    # First, look for explicit "Answer: X" pattern (last occurrence)
-    answer_matches = re.findall(
-        r"[Aa]nswer\s*:\s*\*?\*?\s*\(?([A-Z])\)?", response
-    )
-    if answer_matches:
-        candidate = answer_matches[-1]
-        if candidate in all_choices:
-            return candidate
-
-    for char in (",", ".", "!", "?", ";", ":", "'"):
-        response = response.strip(char)
-    response = " " + response + " "
-
-    candidates: list[str] = []
-    for choice in all_choices:
-        if f"({choice})" in response:
-            candidates.append(choice)
-    if not candidates:
-        for choice in all_choices:
-            if f" {choice} " in response:
-                candidates.append(choice)
-    if not candidates and len(response.split()) > 5:
-        for idx, ans in index2ans.items():
-            if ans and ans.lower() in response.lower():
-                candidates.append(idx)
-    if not candidates:
-        return all_choices[0]
-    if len(candidates) == 1:
-        return candidates[0]
-
-    starts: list[int] = []
-    for can in candidates:
-        pos = response.rfind(f"({can})")
-        if pos == -1:
-            pos = response.rfind(f" {can} ")
-        if pos == -1 and index2ans.get(can):
-            pos = response.lower().rfind(index2ans[can].lower())
-        starts.append(pos)
-    return candidates[int(max(range(len(starts)), key=lambda i: starts[i]))]
