+++ benchmarks/tasks/mmmu.py
@@ -1,0 +1,335 @@
+# SPDX-License-Identifier: Apache-2.0
+"""MMMU benchmark case -- answer parsing, send_fn, metrics, and persistence."""
+
+from __future__ import annotations
+
+import asyncio
+import json
+import logging
+import os
+import re
+import time
+
+import aiohttp
+
+from benchmarks.benchmarker.data import RequestResult
+from benchmarks.benchmarker.runner import SendFn
+from benchmarks.dataset.mmmu import MMMUSample
+
+logger = logging.getLogger(__name__)
+
+SUMMARY_LABEL_WIDTH = 28
+SUMMARY_LINE_WIDTH = 50
+
+
+def _check_is_number(s: str) -> bool:
+    try:
+        float(s.replace(",", ""))
+        return True
+    except Exception:
+        return False
+
+
+def _normalize_str(s: str) -> list:
+    """Normalize a string for open-ended answer comparison."""
+    s = s.strip()
+    if _check_is_number(s):
+        s = s.replace(",", "")
+        try:
+            return [round(float(s), 2)]
+        except Exception:
+            return [s.lower()]
+    return [s.lower()] if len(s) > 1 else [" " + s, s + " "]
+
+
+def _extract_numbers(s: str) -> list[str]:
+    """Extract all numbers (with commas, scientific notation, etc.) from *s*."""
+    pattern_commas = r"-?\b\d{1,3}(?:,\d{3})+\b"
+    pattern_scientific = r"-?\d+(?:\.\d+)?[eE][+-]?\d+"
+    pattern_simple = r"-?(?:\d+\.\d+|\.\d+|\d+\b)(?![eE][+-]?\d+)(?![,\d])"
+    return (
+        re.findall(pattern_commas, s)
+        + re.findall(pattern_scientific, s)
+        + re.findall(pattern_simple, s)
+    )
+
+
+def parse_open_response(response: str) -> list:
+    """Extract answer candidates from an open-ended model response."""
+
+    def _get_key_subresponses(resp: str) -> list[str]:
+        resp = resp.strip().strip(".").lower()
+        subs = re.split(r"\.\s(?=[A-Z])|\n", resp)
+        indicators = [
+            "could be ",
+            "so ",
+            "is ",
+            "thus ",
+            "therefore ",
+            "final ",
+            "answer ",
+            "result ",
+        ]
+        keys: list[str] = []
+        for i, s in enumerate(subs):
+            cands = [*indicators]
+            if i == len(subs) - 1:
+                cands.append("=")
+            shortest = None
+            for ind in cands:
+                if ind in s:
+                    part = s.split(ind)[-1].strip()
+                    if not shortest or len(part) < len(shortest):
+                        shortest = part
+            if shortest and shortest not in (":", ",", ".", "!", "?", ";", "'"):
+                keys.append(shortest)
+        return keys or [resp]
+
+    key_resps = _get_key_subresponses(response)
+    pred_list = key_resps.copy()
+    for r in key_resps:
+        pred_list.extend(_extract_numbers(r))
+    out: list = []
+    for x in pred_list:
+        out.extend(_normalize_str(x))
+    return list(dict.fromkeys(out))
+
+
+def eval_open(gold, preds: list) -> bool:
+    """Check if any prediction matches the gold answer (fuzzy)."""
+    if isinstance(gold, list):
+        norm_answers: list = []
+        for ans in gold:
+            norm_answers.extend(_normalize_str(ans))
+    else:
+        norm_answers = _normalize_str(gold)
+    for p in preds:
+        if isinstance(p, str):
+            for na in norm_answers:
+                if isinstance(na, str) and na in p:
+                    return True
+        else:
+            if p in norm_answers:
+                return True
+    return False
+
+
+def parse_multi_choice_response(
+    response: str,
+    all_choices: list[str],
+    index2ans: dict[str, str],
+) -> str:
+    """Extract a single answer letter from the model response.
+
+    Priority: ``Answer: X`` → ``(A)`` bracket → ``·A·`` space-padded →
+    option-text match → last-occurrence tie-break → fallback to first choice.
+    """
+    answer_matches = re.findall(r"[Aa]nswer\s*:\s*\*?\*?\s*\(?([A-Z])\)?", response)
+    if answer_matches:
+        candidate = answer_matches[-1]
+        if candidate in all_choices:
+            return candidate
+
+    for char in (",", ".", "!", "?", ";", ":", "'"):
+        response = response.strip(char)
+    response = " " + response + " "
+
+    candidates: list[str] = []
+    for choice in all_choices:
+        if f"({choice})" in response:
+            candidates.append(choice)
+    if not candidates:
+        for choice in all_choices:
+            if f" {choice} " in response:
+                candidates.append(choice)
+    if not candidates and len(response.split()) > 5:
+        for idx, ans in index2ans.items():
+            if ans and ans.lower() in response.lower():
+                candidates.append(idx)
+    if not candidates:
+        return all_choices[0]
+    if len(candidates) == 1:
+        return candidates[0]
+
+    starts: list[int] = []
+    for can in candidates:
+        pos = response.rfind(f"({can})")
+        if pos == -1:
+            pos = response.rfind(f" {can} ")
+        if pos == -1 and index2ans.get(can):
+            pos = response.lower().rfind(index2ans[can].lower())
+        starts.append(pos)
+    return candidates[int(max(range(len(starts)), key=lambda i: starts[i]))]
+
+
+def make_mmmu_send_fn(
+    model_name: str,
+    api_url: str,
+    *,
+    max_tokens: int = 2048,
+    temperature: float = 0.0,
+) -> SendFn:
+    """Return a *send_fn* that sends an MMMUSample to /v1/chat/completions.
+
+    Uses the sglang-omni request format with a top-level ``images`` field
+    and ``modalities: ["text"]`` (text-only output, no audio generation).
+    """
+
+    async def send_fn(
+        session: aiohttp.ClientSession, sample: MMMUSample
+    ) -> RequestResult:
+        result = RequestResult(
+            request_id=sample.sample_id,
+            text=sample.prompt[:60],
+        )
+
+        payload = {
+            "model": model_name,
+            "messages": [{"role": "user", "content": sample.prompt}],
+            "images": sample.image_uris,
+            "modalities": ["text"],
+            "max_tokens": max_tokens,
+            "temperature": temperature,
+            "stream": False,
+        }
+
+        start_time = time.perf_counter()
+        try:
+            async with session.post(api_url, json=payload) as response:
+                response.raise_for_status()
+                body = await response.json()
+
+            content = body.get("choices", [{}])[0].get("message", {}).get("content", "")
+            result.text = content or ""
+            result.is_success = True
+
+            usage = body.get("usage", {})
+            if usage:
+                result.prompt_tokens = usage.get("prompt_tokens", 0)
+                result.completion_tokens = usage.get("completion_tokens", 0)
+
+            elapsed = time.perf_counter() - start_time
+            result.engine_time_s = elapsed
+            if result.completion_tokens > 0 and result.engine_time_s > 0:
+                result.tok_per_s = result.completion_tokens / result.engine_time_s
+        except (aiohttp.ClientError, asyncio.TimeoutError) as exc:
+            result.error = str(exc)
+        finally:
+            result.latency_s = time.perf_counter() - start_time
+
+        return result
+
+    return send_fn
+
+
+def compute_mmmu_metrics(
+    samples: list[MMMUSample],
+    results: list[RequestResult],
+) -> tuple[dict, list[dict]]:
+    """Parse answers, compute accuracy, and build per-sample detail list.
+
+    Returns ``(summary_dict, per_sample_list)``.
+    """
+    correct = 0
+    failed = 0
+    per_sample: list[dict] = []
+
+    for sample, result in zip(samples, results):
+        if not result.is_success:
+            per_sample.append(
+                {
+                    "sample_id": sample.sample_id,
+                    "subject": sample.subject,
+                    "question_type": sample.question_type,
+                    "expected": sample.answer,
+                    "predicted": "",
+                    "raw_response": result.error,
+                    "is_correct": False,
+                    "is_success": False,
+                    "error": result.error,
+                    "latency_s": round(result.latency_s, 4),
+                    "prompt_tokens": result.prompt_tokens,
+                    "completion_tokens": result.completion_tokens,
+                    "tok_per_s": (
+                        round(result.tok_per_s, 1) if result.tok_per_s > 0 else None
+                    ),
+                }
+            )
+            failed += 1
+            continue
+
+        gold = sample.answer
+        if (
+            sample.question_type == "multiple-choice"
+            and sample.all_choices
+            and sample.index2ans
+        ):
+            predicted = parse_multi_choice_response(
+                result.text,
+                sample.all_choices,
+                sample.index2ans,
+            )
+            is_correct = gold is not None and predicted == gold
+        else:
+            parsed_list = parse_open_response(result.text)
+            is_correct = gold is not None and eval_open(gold, parsed_list)
+            predicted = ", ".join(map(str, parsed_list))
+
+        if is_correct:
+            correct += 1
+
+        per_sample.append(
+            {
+                "sample_id": sample.sample_id,
+                "subject": sample.subject,
+                "question_type": sample.question_type,
+                "expected": sample.answer,
+                "predicted": predicted,
+                "raw_response": result.text,
+                "is_correct": is_correct,
+                "is_success": True,
+                "error": "",
+                "latency_s": round(result.latency_s, 4),
+                "prompt_tokens": result.prompt_tokens,
+                "completion_tokens": result.completion_tokens,
+                "tok_per_s": (
+                    round(result.tok_per_s, 1) if result.tok_per_s > 0 else None
+                ),
+            }
+        )
+
+    total = len(samples)
+    accuracy = correct / total if total > 0 else 0.0
+
+    summary = {
+        "total_samples": total,
+        "correct": correct,
+        "accuracy": round(accuracy, 4),
+        "failed": failed,
+    }
+    return summary, per_sample
+
+
+def print_mmmu_accuracy_summary(metrics: dict, model_name: str) -> None:
+    """Print formatted MMMU accuracy summary to stdout."""
+    lw = SUMMARY_LABEL_WIDTH
+    print(f"\n{'=' * SUMMARY_LINE_WIDTH}")
+    print(f"  MMMU Accuracy — {model_name}")
+    print(f"{'=' * SUMMARY_LINE_WIDTH}")
+    print(f"  {'Total samples:':<{lw}} {metrics['total_samples']}")
+    print(f"  {'Correct:':<{lw}} {metrics['correct']}")
+    print(
+        f"  {'Accuracy:':<{lw}} {metrics['accuracy']:.4f} "
+        f"({metrics['accuracy'] * 100:.1f}%)"
+    )
+    print(f"  {'Failed requests:':<{lw}} {metrics['failed']}")
+    print(f"{'=' * SUMMARY_LINE_WIDTH}\n")
+
+
+def save_mmmu_results(results: dict, output_dir: str) -> None:
+    """Save results to ``mmmu_results.json`` in *output_dir*."""
+    os.makedirs(output_dir, exist_ok=True)
+    path = os.path.join(output_dir, "mmmu_results.json")
+    with open(path, "w") as f:
+        json.dump(results, f, indent=2, ensure_ascii=False)
+    logger.info("Results saved to %s", path)
